Move web3connect files to the front side
