+++ src/front/client/index.html
@@ -824,6 +824,15 @@
       window.widgetName = 'own widget name';
       window.buyViaCreditCardLink = 'https://buy.itez.com/'; // need replace at 'https://buy.itez.com/swaponline' by default
 
+
+
+      // window.CUR_ETH_DISABLED = true
+      // window.CUR_BNB_DISABLED = true
+      // window.CUR_MATIC_DISABLED = true
+      // window.CUR_ARBITRUM_DISABLED = true
+
+
+
       if (window.brandColor) {
         document.documentElement.style.setProperty("--color-brand", window.brandColor);
       }
