+++ src/front/shared/components/modals/ConnectWalletModal/ConnectWalletModal.tsx
@@ -6,7 +6,7 @@ import cssModules from 'react-css-modules'
 import cx from 'classnames'
 import styles from './ConnectWalletModal.scss'
 import { externalConfig } from 'helpers'
-import SUPPORTED_PROVIDERS from 'common/web3connect/providers/supported'
+import SUPPORTED_PROVIDERS from 'helpers/web3connect/providers/supported'
 import { constants, links, metamask } from 'helpers'
 import { localisedUrl } from 'helpers/locale'
 import { Button } from 'components/controls'
