+++ src/front/shared/helpers/web3connect/providers/InjectedProvider.ts
