+++ src/front/shared/helpers/web3connect/providers/InjectedType.ts
