+++ src/front/shared/helpers/web3connect/providers/WalletConnectProvider.ts
